+++ app/build.gradle
@@ -56,7 +56,6 @@ dependencies {
     implementation 'com.android.support.test:rules:0.5'
     implementation 'com.android.support.test.uiautomator:uiautomator-v18:2.1.2'
 
-
     implementation('com.zarinpal:purchase:0.0.8-beta') {
         exclude group: 'com.mcxiaoke.volley',
                 module: 'library'
