+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,6 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#F9FBFF"
     app:layout_anchor="@+id/constraintLayout"
     app:layout_anchorGravity="center">
