+++ app/src/main/res/layout/fragment_shopping.xml
@@ -8,12 +8,15 @@
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/ShoppingRecycler"
-        android:layout_width="0dp"
-        android:layout_height="680dp"
+        android:layout_width="413dp"
+        android:layout_height="652dp"
+        app:layout_constraintBottom_toTopOf="@+id/cardN"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/cardN" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0"
+        tools:ignore="MissingConstraints" />
 
     <android.support.v7.widget.CardView
         android:id="@+id/cardN"
