+++ build.gradle
@@ -7,11 +7,11 @@ buildscript {
         google()
         //maven()
         jcenter()
-        //maven { url "https://s3.amazonaws.com/repo.commonsware.com" }
+        maven { url "https://s3.amazonaws.com/repo.commonsware.com" }
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:3.5.3'
